+++ Assets/Scripts/BaseBehaviour.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1145297d440216046932fa4fc7431839
+timeCreated: 1473361194
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
