+++ Assets/Scripts/Bixao.cs
@@ -1,38 +1,23 @@
 ﻿using UnityEngine;
 using System.Collections;
 
-public class Bixao : MonoBehaviour {
-
-    private Rigidbody2D rb;
-    private bool becameVisible = false;
+public class Bixao : BaseBehaviour {
 
     // Use this for initialization
-    void Start () {
-        rb = GetComponent<Rigidbody2D>();
+    public override void Start()
+    {
+        base.Start();
     }
-	
-	// Update is called once per frame
-	void Update () {
-        rb.velocity = new Vector2(rb.velocity.x, -30f);
 
-        if (becameVisible)
-            IsVisible();
+    // Update is called once per frame
+    public override void Update()
+    {
+        base.rb.velocity = new Vector2(base.rb.velocity.x, -30f);
     }
 
     void OnTriggerEnter2D(Collider2D col)
     {
         if (col.gameObject.tag == "Player")
             Destroy(col.gameObject);
-    }
-
-    void IsVisible()
-    {
-        if (!GetComponent<Renderer>().isVisible)
-            Destroy(this.gameObject);
-    }
-
-    void OnBecameVisible()
-    {
-        becameVisible = true;
     }
 }
