+++ Assets/Scripts/Bomb.cs
@@ -1,25 +1,20 @@
 ﻿using UnityEngine;
 using System.Collections;
 
-public class Bomb : MonoBehaviour {
+public class Bomb : BaseBehaviour {
 
     public GameObject bombExplosion;
-    private Rigidbody2D rb;
-    private MainScript mainScript;
-    private bool becameVisible = false;
 
     // Use this for initialization
-    void Start() {
-        rb = GetComponent<Rigidbody2D>();
-        mainScript = (MainScript)FindObjectOfType(typeof(MainScript));
+    public override void Start()
+    {
+        base.Start();
     }
 
     // Update is called once per frame
-    void Update() {
-        rb.velocity = new Vector2(-1 * mainScript.speed, rb.velocity.y);
-
-        if (becameVisible)
-            IsVisible();
+    public override void Update()
+    {
+        base.Update();
     }
 
     void OnTriggerEnter2D(Collider2D col)
@@ -40,16 +35,5 @@ public class Bomb : MonoBehaviour {
     {
         Instantiate(bombExplosion, transform.position, transform.rotation);
         Destroy(gameObject);
-    }
-
-    void IsVisible()
-    {
-        if (!GetComponent<Renderer>().isVisible)
-            Destroy(this.gameObject);
-    }
-
-    void OnBecameVisible()
-    {
-        becameVisible = true;
     }
 }
