+++ Assets/Scripts/Coin.cs
@@ -1,26 +1,21 @@
 ﻿using UnityEngine;
 using System.Collections;
 
-public class Coin : MonoBehaviour {
+public class Coin : BaseBehaviour {
 
-    private MainScript mainScript;
-    private Rigidbody2D rb;
     private Player player;
-    private bool becameVisible = false;
 
     // Use this for initialization
-    void Start () {
-        mainScript = (MainScript)FindObjectOfType(typeof(MainScript));
-        rb = GetComponent<Rigidbody2D>();
+    public override void Start()
+    {
+        base.Start();
         player = (Player)FindObjectOfType(typeof(Player));
     }
-	
-	// Update is called once per frame
-	void Update () {
-        rb.velocity = new Vector2(-1 * mainScript.speed, rb.velocity.y);
 
-        if (becameVisible)
-            IsVisible();
+    // Update is called once per frame
+    public override void Update()
+    {
+        base.Update();
     }
 
     void OnTriggerEnter2D(Collider2D col)
@@ -31,16 +26,5 @@ public class Coin : MonoBehaviour {
             player.coinsCollected += 1;
             print(player.coinsCollected);
         }
-    }
-
-    void IsVisible()
-    {
-        if (!GetComponent<Renderer>().isVisible)
-            Destroy(this.gameObject);
-    }
-
-    void OnBecameVisible()
-    {
-        becameVisible = true;
     }
 }
